Fix arithmetic overflow when calculating PokerHand value

+++ src/Blef.GameLogic/PokerHands/Flush.cs
@@ -14,6 +14,6 @@
             return table.HasSuit(suit);
         }
 
-        protected override int Value => 1000000 * GetSuitValue(suit);
+        protected override long Value => 1000000 * GetSuitValue(suit);
     }
 }
