+++ src/Blef.GameLogic/PokerHands/HighCard.cs
@@ -15,6 +15,6 @@ namespace Blef.GameLogic.PokerHands
             return table.HasRank(rank);
         }
 
-        protected override int Value => GetRankValue(rank);
+        protected override long Value => GetRankValue(rank);
     }
 }
